+++ src/SystemJournal/DTO/JournalItemCollection.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B24io\Loyalty\SDK\SystemJournal\DTO;
+
+/**
+ * JournalItemCollection collection
+ *
+ * @method attach(JournalItem $journalItem)
+ * @method JournalItem current()
+ */
+class JournalItemCollection extends \SplObjectStorage
+{
+}